Add screenshot generation to build process
Adds a 'takescreenshots' command which will scan for test_screenshots.py
files and run the testcases inside. These should be based on the included
LiveServerTestCase subclass which uses phantomjs to generate screenshots of
the pages navigated to by the test methods.

The LiveServerThread used by the testcase can be replaced with a subclass
which allows for mocking of api responses etc