Add support logic for the Cloud Platform migration
This adds a new middleware that overrides the default request
handlers and returns a maintenance page if
settings.CLOUD_PLATFORM_MIGRATION_MODE = 'maintenance' or a page
saying that we have moved to a different domain if
settings.CLOUD_PLATFORM_MIGRATION_MODE = 'moved'.

The template for the 'we have moved' mode, automatically redirects
to the new url unless the browser is IE8.