nomis: Retry twice for nomis.get_* requests
Some of the requests to NOMIS API are timing out despite the timeout being
15s (which is high).

The `retries` parameter for the following functions now defaults to `2`:
- `nomis.get_account_balances()`
- `nomis.get_transaction_history()`
- `nomis.get_photograph_data()`
- `nomis.get_location()`

Hopefully this may help with some of the requests failing.

NOTE: There may be other places where HTTP requests are made and these
may still timeout but this is a first step and it should be relatively
safe.

Ticket: https://dsdmoj.atlassian.net/browse/MTP-1768