RFC: Retry mechanism for auth/api_client
This re-uses the `nomis.Retry` class/logic and do **not** retry by default.

As retrying is by disabled by default disabled (similarly to default in `nomis` module),
this means that the code using this class/module will have to explicitly pass `retries`
if it wants to retry on failure.

Unfortunately these requests may be in other apps using `mtp_common`, e.g. `bank-admin`
and the fact that these requests use `mtp_common.auth` may not be immediately obvious,
For example, in this [`bank_admin.views.get_missing_downloads()`], the code calls
`api_session.get()`, in this case `api_session` is an instance of
`mtp_common.auth.api_client.MoJOAuth2Session` returned by
`mtp_common.auth.api_client.get_api_session()` but it requires a bit of following
the breadcrums or intuition to find out.

Above may be mitigated by defaulting to `2` retries, at least for `GET` requests,
but this seems a bit of a nuclear option and may have unexpected consequences.

NOTE: This logic is agnostic of the request method, e.g. potentially the client code may
want to retry a `PATCH` request if it thinks it's safe, e.g. it has knowledge that specific
API request is idempotent for example.

Related to ticket: https://dsdmoj.atlassian.net/browse/MTP-1768

[`bank_admin.views.get_missing_downloads()`]: https://github.com/ministryofjustice/money-to-prisoners-bank-admin/blob/2ceae3f4fa9a6992bf505f88585f3682c0079afd/mtp_bank_admin/apps/bank_admin/views.py#L37-L44