Added 'account requests' banner as per Justin prototype
Prototype for MTP-1818 (https://pm-intelligence-tool.herokuapp.com/)
has a banner which is similar to the one in `cashbook` but adds a link to
manage these requests.

This makes sense and we should use the same banner for both `cashbook`
and `noms-ops` so it makes sense to move this into `mtp-common` and
re-use in both places.

MTP-1818: https://dsdmoj.atlassian.net/browse/MTP-1818