Bumped transifex-client to support Python 3.9
When trying to install `mtp-common` locally on a machine with Python 3.9
it fails because of the 'transifex-client>=0.13,<0.14' constrain.

Latest transifex-client<0.14 (==0.13.12) requires python<3.9:

https://pypi.org/project/transifex-client/0.13.12/

transifex-client==0.14.2 supports Python 3.9 (and also 3.6)

https://pypi.org/project/transifex-client/0.14.2/

We currently don't use Python 3.9 on production but locally editors like
VS Code are relying on `mtp-common` to be correctly installed in the
virtual environment in order to be useful (e.g. to jump to code definition
etc).