bumped to 12.0.1: Node.js 16
This should not be a breaking change as we're mainly using Node.js to
build assets and that seem to work as before with Node.js 16.

Other packages don't seem to be particularly bothered by the version of
Node.js and the Node.js changelog didn't seem to highlight any breaking
change that should affect us, so only bumping the patch number to keep
things simple.