MTP-1937: Analytics.js: Send GA4 events for `data-analytics`'s clicks
On `data-analytics`' clicks we're sending page view/custom events to GA4.

This was done already for the legacy GA. These attributes have a
comma-separated value with the 1st part determining whether the click
has to trigger a page view or a custom event, e.g.
- `data-analytics="pageview,/batch/-dialog_close/"`
- `data-analytics="event,PrisonConfirmation,Add,{{ form.all_prisons_code }}"`

The new event handler does the same thing but using the new GA4' `gtag()`
function.

NOTE: The existing markup doesn't change for simplicity sake but in the
      case of GA4 a standard `page_view` event is sent (not `pageview`)