Bump package version
When I upgraded the version of Twine to solve the
release issues I then made a release via GH release
tag...however I didn't bump the version of the package
in the code. So right now latest version on PyPi is
`19.0.1` while on GH releases latest is `19.0.2`.

Not a big deal, it will realign next time we bump
the version (to `19.0.3` or whatever)